# apps/web/public/openapi.yaml
openapi: "3.0.3"
info:
  title: BookKeep API
  description: REST API for BookKeep — bookkeeping, reconciliation, reports, and team management.
  version: "1.0.0"
  contact:
    email: support@bookkeep.app
servers:
  - url: https://bookkeep.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Development
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Account:
      type: object
      properties:
        _id:
          type: string
        code:
          type: string
          example: "1000"
        name:
          type: string
          example: "Cash"
        type:
          type: string
          enum: [asset, liability, equity, revenue, expense]
        normalBalance:
          type: string
          enum: [debit, credit]
        isActive:
          type: boolean
    JournalLine:
      type: object
      properties:
        accountCode:
          type: string
        description:
          type: string
        debit:
          type: number
        credit:
          type: number
    JournalEntry:
      type: object
      properties:
        _id:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/JournalLine"
...
paths:
  /accounts:
    get:
      summary: List all active accounts
      tags: [Accounts]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Account"
    post:
      summary: Create a new account
      tags: [Accounts]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, type]
              properties:
                code:
                  type: string
                  example: "1100"
                name:
                  type: string
                  example: "Accounts Receivable"
                type:
                  type: string
                  enum: [asset, liability, equity, revenue, expense]
                normalBalance:
                  type: string
                  enum: [debit, credit]
      responses:
        "201":
          description: Account created
        "400":
          description: Missing required fields
        "409":
          description: Account code already exists
...
  /entries/{entryId}:
    get:
      summary: Get a single journal entry
      tags: [Entries]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: entryId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entry with lines
    put:
      summary: Update a journal entry
      tags: [Entries]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntry"
      responses:
        "200":
          description: Entry updated
    delete:
      summary: Soft-delete a journal entry
      tags: [Entries]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Entry moved to trash
...